dock-server/app/requirements.txt
pandas
sklearn
numpy
mysql-connector-python
tensorflow
